==> src/analysis/s1_load-from-dbs.ipynb <==
# Load data from databases
This notebook uses the below two prepared databases converted from the raw records.
The output is a flat and de-normalised table.

1 data/accidents_shenzhen.sqlite

2 data/accidents_shenzhen_GIS.sqlite (crash geolocations by querying Google API)

output: data/data_s1.csv

%load_ext autoreload
%autoreload 2

import sqlite3
import pandas as pd
import numpy as np
import os
from pathlib import Path
import json

data_path = os.path.join(Path(os.getcwd()).parent.parent, 'data/')
with open(os.path.join(data_path, 'value_mapping.txt'), encoding='utf-8') as json_file:
    value_dict = json.load(json_file)

## 1 Load data from database for merging and preprocessing

db = "accidents_shenzhen.sqlite"
db_loc = "accidents_shenzhen_GIS.sqlite"

### 1.1 Accident records

## Load accident records
conn = sqlite3.connect(os.path.join(data_path, db))
query = 'SELECT id, acc_id, time, type, weather, reason FROM records'
df_rec = pd.read_sql_query(query, conn)

# Convert acc_id
df_rec.acc_id = df_rec.acc_id.apply(lambda x: int.from_bytes(x, byteorder='little'))
## Preprocess df_rec
df_rec.fillna(value=pd.np.nan, inplace=True)
value_dict_rev = {v: k for k, v in value_dict['acc_type'].items()}
value_dict_rev["滚翻"] = "32"
df_rec.loc[:, 'type'] = df_rec.loc[:, 'type'].apply(lambda x: value_dict_rev[x] if x in value_dict_rev else x)
convert_dict = {'id': pd.Int64Dtype(), 'type': pd.Int64Dtype(), 'reason': pd.Int64Dtype()}
df_rec.loc[:, 'weather'] = df_rec.loc[:, 'weather'].apply(lambda x: int.from_bytes(x, byteorder='little'))
df_rec.loc[:, 'type'] = df_rec.loc[:, 'type'].apply(lambda x: float(x))
df_rec.loc[:, 'reason'] = df_rec.loc[:, 'reason'].apply(lambda x: float(x))
df_rec = df_rec.astype(convert_dict)
df_rec.head()

### 1.2 Accident locations

## Load accident locations
conn = sqlite3.connect(os.path.join(data_path, db_loc))
query = 'SELECT rec_id, lat, long FROM gis'
df_loc= pd.read_sql_query(query, conn)
convert_dict = {'rec_id': int, 'lat': float, 'long': float}
df_loc = df_loc.astype(convert_dict)
df_loc.loc[:, 'lat'] = df_loc.loc[:, 'lat'].apply(lambda x: np.nan if x == 999 else x)
df_loc.loc[:, 'long'] = df_loc.loc[:, 'long'].apply(lambda x: np.nan if x == 999 else x)
df_loc.head()

### 1.3 Accident severity

## Load accident severity
conn = sqlite3.connect(os.path.join(data_path, db))
query = 'SELECT rec_id, death, death_s, miss, inj_b, inj_l, inj, loss FROM severity'
df_sev = pd.read_sql_query(query, conn)
for var in ['death', 'death_s', 'miss', 'inj_b', 'inj_l', 'inj', 'loss']:
    df_sev.loc[:, var] = df_sev.loc[:, var].apply(lambda x: int.from_bytes(x, byteorder='little'))
df_sev = df_sev.astype({'rec_id': int})
df_sev.head()

### 1.4 geo

## Load geo
conn = sqlite3.connect(os.path.join(data_path, db))
query = 'SELECT rec_id, road_id FROM geo'
df_geo= pd.read_sql_query(query, conn)
df_geo.fillna(value=pd.np.nan, inplace=True)
df_geo = df_geo.astype({'road_id': float})
df_geo = df_geo.astype({'road_id': pd.Int64Dtype()})

### 1.5 road

## Load road
conn = sqlite3.connect(os.path.join(data_path, db))
query = 'SELECT id, name, type, distr FROM road'
df_road= pd.read_sql_query(query, conn)
df_road.fillna(value=pd.np.nan, inplace=True)
value_dict_rev = {v: k for k, v in value_dict['road_type'].items()}
df_road.loc[:, 'type'] = df_road.loc[:, 'type'].apply(lambda x: value_dict_rev[x] if x in value_dict_rev else x)
convert_dict = {'id': pd.Int64Dtype(), 'type': pd.Int64Dtype(), 'distr': pd.Int64Dtype()}
df_road.loc[:, 'type'] = df_road.loc[:, 'type'].apply(lambda x: float(x))
df_road.loc[:, 'distr'] = df_road.loc[:, 'distr'].apply(lambda x: float(x))
df_road = df_road.astype(convert_dict)

## Join road info to record
df_road = pd.merge(df_geo, df_road, left_on='road_id', right_on='id').drop(columns=['road_id', 'id'])
df_road.rename(columns={'name': 'road_name', 'type': 'road_type'}, inplace=True)
df_road.head()


### 1.6 People involved in the accidents
This step takes all the crash participants into consideration.

# Read this part of data directly from source files
var_list = ['事故编号', '性别', '年龄', '文化程度', '事故责任', '车辆类型', '交通方式']
df_ppl_list = []
for year in [2014,2015,2016]: #,',2015,2016'
    for selec in ['上半年','下半年']: #,'下半年'
        basicName = '人员信息表'
        xl = pd.ExcelFile(os.path.join(data_path, str(year)+'年'+basicName+'.xls'))
        df_pl = xl.parse(selec)
        df_pl = df_pl.loc[:, var_list]
        df_pl.columns = ['acc_id', 'gender', 'age', 'edu', 'respon', 'veh_type', 'travel_mode']
        df_pl = df_pl.loc[df_pl.acc_id.isin(df_rec.acc_id), :]
        df_ppl_list.append(df_pl)
df_ppl = pd.concat(df_ppl_list)

# Complete veh_type and travel_mode with the other
df_ppl['veh_type'].fillna(df_ppl['travel_mode'], inplace=True)
df_ppl['travel_mode'].fillna(df_ppl['veh_type'], inplace=True)
df_ppl.head()

## 2 Combine data together for further modelling

## Merge location
df = pd.merge(df_rec, df_loc, left_on='id', right_on='rec_id', how="outer").drop(columns=['rec_id'])
print('Merged records and locations: %s' % len(df))

df = pd.merge(df, df_road, left_on='id', right_on='rec_id', how="outer").drop(columns=['rec_id'])
print('Merged road info: %s' % len(df))

df = pd.merge(df, df_sev, left_on='id', right_on='rec_id', how="outer").drop(columns=['rec_id'])
print('Merged severity info: %s' % len(df))

df = pd.merge(df, df_ppl, left_on='acc_id', right_on='acc_id', how="left")
print('Merged people involved: %s' % len(df))

df.drop(columns=['loss'], inplace=True)
print(df.info())

df.head()

df.to_csv(os.path.join(data_path, "data_s1.csv"), index=False)
