==> src/analysis/s2_preprocessing.ipynb <==
# Data preprocessing

Input: data/data_s1.csv

Process each variable and clean up.

Output: data/data_s2.csv

%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np
import os
from pathlib import Path
import json
from datetime import datetime
import matplotlib.pyplot as plt


data_path = os.path.join(Path(os.getcwd()).parent.parent, 'data/')
with open(os.path.join(data_path, 'value_mapping.txt'), encoding='utf-8') as json_file:
    value_dict = json.load(json_file)

# Load data
df = pd.read_csv(os.path.join(data_path, "data_s1.csv"))
# Drop the below columns for being out of interest
df.drop(columns=['road_name', 'distr', 'miss', 'acc_id'], inplace=True)
df.info()

## 1 Process gender, edu, veh_type, deaths & injuries, and time
data/value_mapping.txt

### 1.1 Prepare for veh_type mapping rules

df_veh_type = pd.DataFrame(list(df.veh_type.unique()), columns=['veh_type'])
df_veh_type.dropna(inplace=True)
def cate_func(x):
    if '轿车' in x:
        return 1
    elif ('货车' in x) or ('牵引' in x) or ('挂车' in x) or ('厢车' in x):
        return 3
    elif '客车' in x:
        return 2
    elif '摩托车' in x:
        return 4
    elif 'F' in x:
        return 5
    else:
        return 6

df_veh_type.loc[:, 'cate'] = df_veh_type.loc[:, 'veh_type'].apply(lambda x: cate_func(x) if x != np.nan else x)
veh_type_dict = {row['veh_type']: row['cate'] for _, row in df_veh_type.iterrows()}

### 1.2 Prepare for travel_mode mapping rules

df_travel_mode = pd.DataFrame(list(df.travel_mode.unique()), columns=['travel_mode'])
df_travel_mode.dropna(inplace=True)
def cate_func_m(x):
    if '步行' in x:
        return 1
    elif ('三轮车' in x) or ('自行车' in x) or ('手推车' in x) or ('电动自行车' in x) or ('其它非机动车' in x):
        return 2
    elif '其它' in x:
        return 4
    else:
        return 3
df_travel_mode.loc[:, 'cate'] = df_travel_mode.loc[:, 'travel_mode'].apply(lambda x: cate_func_m(x) if x != np.nan else x)
travel_mode_dict = {row['travel_mode']: row['cate'] for _, row in df_travel_mode.iterrows()}

### 1.3 Process gender, edu, respon, and veh_type
gender, edu, respon from externally defined mapping rules: value_dict.

# Driver gender
df.loc[:, 'gender'] = df.loc[:, 'gender'].apply(lambda x: str(int(value_dict['gender'][x])) if x in value_dict['gender'] else np.nan)
df.loc[:, 'gender'].fillna("Unknown", inplace=True)

# Driver edu
df.loc[:, 'edu'] = df.loc[:, 'edu'].apply(lambda x: str(int(value_dict['edu'][x])) if x in value_dict['edu'] else np.nan)
df.loc[:, 'edu'].fillna("Unknown", inplace=True)

# Driver respon
df.loc[:, 'respon'] = df.loc[:, 'respon'].apply(lambda x: str(int(value_dict['respon'][x])) if x in value_dict['respon'] else np.nan)
df.loc[:, 'respon'].fillna("Unknown", inplace=True)

# Driver vehicle type
df.loc[:, 'veh_type'] = df.loc[:, 'veh_type'].apply(lambda x: str(int(veh_type_dict[x])) if x in veh_type_dict else np.nan)
df.loc[:, 'veh_type'].fillna("Unknown", inplace=True)

# Driver vehicle type
df.loc[:, 'travel_mode'] = df.loc[:, 'travel_mode'].apply(lambda x: str(int(travel_mode_dict[x])) if x in travel_mode_dict else np.nan)
df.loc[:, 'travel_mode'].fillna("Unknown", inplace=True)

# convert_dict = {'gender': pd.Int64Dtype(), 'edu': pd.Int64Dtype(), 'veh_type': pd.Int64Dtype()}
# df = df.astype(convert_dict)
for var in ['gender', 'edu', 'veh_type', 'respon', 'travel_mode']:
    print(var, df[var].unique(), '\n')

### 1.4 Merge deaths and injuries

df.loc[:, 'injs'] = df.loc[:, ['inj', 'inj_b', 'inj_l']].sum(axis=1)
df.loc[:, 'deaths'] = df.loc[:, ['death', 'death_s']].sum(axis=1)
df.drop(columns=['inj', 'inj_b', 'inj_l', 'death', 'death_s'], inplace=True)
df.iloc[0:3].transpose()

### 1.5 Process time to get weekday (0 for weekend, 1 for weekday), time of day

df.loc[:, 'time'] = df.loc[:, 'time'].apply(lambda x: datetime.strptime(x.strip(), '%Y-%m-%d %H:%M:%S'))
df.loc[:, 'weekday'] = df.loc[:, 'time'].apply(lambda x: x.weekday() + 1)
df.loc[:, 'hour'] = df.loc[:, 'time'].apply(lambda x: x.hour)
df.drop(columns=['time'], inplace=True)
df.iloc[0:3].transpose()

## 2 Clean up the categories for further analysis
### 2.1 age
Break it into age groups and label missing fields as Unknown.

df.loc[:, 'age'] = pd.cut(df.age, bins=[0, 18, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 150])
age_cat_dict = {'(0, 18]': '< 18', '(70, 150]': '> 70'}
df.loc[:, 'age'] = df.loc[:, 'age'].apply(lambda x: age_cat_dict[str(x)] if str(x) in age_cat_dict else str(x))
df.loc[:, 'age'].cat.add_categories("Unknown", inplace=True)
df.loc[:, 'age'].fillna("Unknown", inplace=True)
print('age', df['age'].unique(), '\n')

### 2.2 reason
Select top reasons, set the rest as Others, and label missing fields as Unknown.

## Merge two categories that are essentially the same (1301, 2004 -> 2004)
df.loc[:, 'reason'] = df.loc[:, 'reason'].apply(lambda x: 2004 if x == 1301 else x)

# Check the top 20 reasons
acc_reason = df.drop_duplicates(subset=["id"]).groupby('reason')[['id']].count()/len(df.drop_duplicates(subset=["id"]))*100
acc_reason = acc_reason.reset_index().sort_values(by='id', ascending=False).rename(columns={'id': 'freq'})
acc_reason.loc[:, 'cum_freq'] = acc_reason.loc[:, 'freq'].cumsum()

# Set the top 18 reasons and set nan to Unknown
reason_list = [int(x) for x in acc_reason.head(20).reason if x not in (9901, 9009)]
df.loc[:, 'reason'].fillna("Unknown", inplace=True)

# Convert the rest into str and set the rest as others
df.loc[df['reason'].isin(reason_list), 'reason'] = df.loc[df['reason'].isin(reason_list), 'reason'].apply(lambda x: str(int(x)))
reason_list = [str(int(x)) for x in reason_list] + ['Unknown']
df.loc[~df['reason'].isin(reason_list), 'reason'] = 'Others'

print('reason', df['reason'].unique(), '\n')

### 2.3 type
1. Merge type 31 and 32 into one, coded as 312 (Rollover)

2. Merge type 22 and 23 into one, coded as 223 (Crushing pedestrians)

3. Selected types: 11, 35, 19, 21, 12, 36, 312, 223, 38, 999, 29, 34, 33. Set the rest as Others.
Missing values set to Unknown.

# Check the top 20 types
acc_type = df.drop_duplicates(subset=["id"]).groupby('type')[['id']].count()/len(df)*100
acc_type = acc_type.reset_index().sort_values(by='id', ascending=False).rename(columns={'id': 'freq'})
acc_type.loc[:, 'cum_freq'] = acc_type.loc[:, 'freq'].cumsum()
acc_type.loc[:, 'type_name'] = acc_type.loc[:, 'type'].apply(lambda x: value_dict['acc_type'][str(int(x))] if str(int(x)) in value_dict['acc_type'] else 'unknown')
acc_type.head(30)


# 1 Merge type 31 and 32 into 312
df.loc[:, 'type'] = df.loc[:, 'type'].apply(lambda x: 312 if x in (31, 32) else x)

# 2 Merge type 22 and 23 into 223
df.loc[:, 'type'] = df.loc[:, 'type'].apply(lambda x: 223 if x in (22, 23) else x)

# 3 Select a few types
df.loc[:, 'type'].fillna("Unknown", inplace=True)
type_list = [11, 35, 19, 21, 12, 36, 312, 223, 38, 999, 29, 34, 33]
df.loc[df.type != 'Unknown', 'type'] = df.loc[df.type != 'Unknown', 'type'].apply(lambda x: str(int(x)) if x in type_list else "Others")

print('type', df['type'].unique(), '\n')

### 2.4 injs
Break it into groups and label missing fields as Unknown.

df.loc[:, 'injs_num'] = df.loc[:, 'injs'].copy()
df.loc[:, 'injs'] = pd.cut(df.injs, bins=[-1, 0, 4, 9, 100])
injs_cat_dict = {'(-1, 0]': '= 0', '(9, 100]': '> 9'}
df.loc[:, 'injs'] = df.loc[:, 'injs'].apply(lambda x: injs_cat_dict[str(x)] if str(x) in injs_cat_dict else str(x))
df.loc[:, 'injs'].cat.add_categories("Unknown", inplace=True)
df.loc[:, 'injs'].fillna("Unknown", inplace=True)

### 2.5 deaths
Break it into groups and label missing fields as Unknown.

df.loc[:, 'deaths'].min(), df.loc[:, 'deaths'].max()
df.loc[:, 'deaths_num'] = df.loc[:, 'deaths'].copy()

df.loc[:, 'deaths'] = pd.cut(df.deaths, bins=[-1, 0, 4, 9, 100])
deaths_cat_dict = {'(-1, 0]': '= 0', '(9, 100]': '> 9'}
df.loc[:, 'deaths'] = df.loc[:, 'deaths'].apply(lambda x: deaths_cat_dict[str(x)] if str(x) in deaths_cat_dict else str(x))
df.loc[:, 'deaths'].cat.add_categories("Unknown", inplace=True)
df.loc[:, 'deaths'].fillna("Unknown", inplace=True)

### 2.6 road_type

df.loc[:, 'road_type'].fillna("Unknown", inplace=True)
df.loc[:, 'road_type'] = df.loc[:, 'road_type'].apply(lambda x: str(int(x)) if x != "Unknown" else x)

### 2.7 weather
1. Merge 8 and 9 into 89 (Haze or fog).

2. Remove 5 and 6 because there is no way Shenzhen will observe a weather like snow and blizzard. Label them as Unknown.

df.loc[:, 'weather'].fillna("Unknown", inplace=True)
df.loc[:, 'weather'] = df.loc[:, 'weather'].apply(lambda x: "89" if x in [8, 9] else x)
df.loc[:, 'weather'] = df.loc[:, 'weather'].apply(lambda x: "Unknown" if x in [5, 6] else x)

df.loc[df.weather.isin([1,2,3,4]), 'weather'] = df.loc[df.weather.isin([1,2,3,4]), 'weather'].apply(lambda x: str(int(x)))

print('weather', df['weather'].unique(), '\n')

# Save data
df.to_csv(os.path.join(data_path, 'data_s2.csv'), index=False, encoding='utf-8-sig')
