==> src/analysis/s4_descriptive_analysis.ipynb <==
# Descriptive analysis

Input: data/data_s3.csv

1. Translate (map) numerical levels of variables into readable English version.

2. Generate statistics tables for all the crash records.

3. Generate statistics tables for all the crash records' participants.

4. Process to generate data for Bayesian network analysis.

Output: data/data_s4.csv

%load_ext autoreload
%autoreload 2

import pandas as pd
import scipy.stats as stats
import os
from pathlib import Path
import json
import numpy as np
import matplotlib.pyplot as plt

figure_path = os.path.join(Path(os.getcwd()).parent.parent, 'figures/')
data_path = os.path.join(Path(os.getcwd()).parent.parent, 'data/')
with open(os.path.join(data_path, 'value_mapping.txt'), encoding='utf-8') as json_file:
    value_dict = json.load(json_file)

# Load data
df = pd.read_csv(os.path.join(data_path, 'data_s3.csv'), low_memory=False)

## 1 Data summary: variables and levels
### 1.1 Summarize variables and its levels

# Rename columns
var_mapping = {'weekday': 'Day of week',
               'hour': 'Time of day',
               'weather': 'Weather',
               'cluster': 'Land-use cluster',
               'type': 'Crash type',
               'reason': 'Crash causation',
               'road_type': 'Road type',
               'veh_type': 'Vehicle type',
               'injs': 'Injuries (bin)',
               'deaths': 'Deaths (bin)',
               'injs_num': 'Injuries',
               'deaths_num': 'Deaths',
               'age': 'Age',
               'edu': 'Education level',
               'gender': 'Gender',
               'respon': 'Responsibility',
               'travel_mode': 'Travel method'}

df = df.rename(columns={k: v for k, v in var_mapping.items()})

# Extract levels
col_list = list(df.columns)
col_list.remove('id')
level_df_list = []
for var in col_list:
    level_df_list.append(pd.DataFrame([(var, x) for x in df[var].unique()], columns=['var', 'level']))
df_level = pd.concat(level_df_list)
df_level.to_clipboard(index=False)
df.iloc[0]

### 1.2 Manually define the translate of levels for each variable
The file is stored in bn_var_level_translate.xlsx.


## Prepare table for write-up
df_var_list = []
df_lvs = pd.read_excel(os.path.join(Path(os.getcwd()).parent.parent, 'docs/var_level_translate.xlsx'))

for var, frame in df_lvs.groupby('var'):
    ld = '; '.join([' '.join([str(row['level_index']) + ')', str(row['en'])]) for _, row in frame.sort_values(by='level_index').iterrows()])
    df_var_list.append(pd.DataFrame([(var, str(len(frame)), ld)], columns=('Variable', 'Level number', 'Level description')))
df_var_tb = pd.concat(df_var_list)
df_var_tb.to_clipboard(index=False)

## 2 Replace numerical levels with English abbreviations for easy description

df_lvs = df_lvs.astype(str)

for var in df.columns:
    if var not in ['id', 'Injuries', 'Deaths']:
        df.loc[:, var] = df.loc[:, var].astype(str)
        lvs_dict = {row['level']: row['abbr'] for _, row in df_lvs.loc[df_lvs['var'] == var, :].iterrows()}
        df.loc[:, var] = df.loc[:, var].apply(lambda x: lvs_dict[x])

for var in df.columns:
    print(var, df[var].unique(), '\n')

# Remove duplicated records to look at the crash attributes
df_rec = df.drop_duplicates(subset=['id'])


## 3 Statistics of different crash attributes
This part of the results are summarised in docs/article_tables.xlsx.

print(f'Total number of crashes: {len(df_rec)}')
print(f'Total number of the involved traffic participants: {len(df)}')

def basic(data):
    data = pd.merge(data, df_lvs.loc[df_lvs['var'] == var, ['en', 'abbr']],
                    left_on=var, right_on='abbr')
    num = len(data)
    injs = data['Injuries'].sum()
    deaths = data['Deaths'].sum()
    return pd.Series({'Var en': data.iloc[0]['en'],
                      'Crash number': num, 'Crash %': num / len(df_rec) * 100,
                      'Deaths number': deaths, 'Deaths %': deaths / df_rec['Deaths'].sum() * 100,
                      'Injuries number': injs, 'Injuries %': injs / df_rec['Injuries'].sum() * 100
                      })

### 3.1 Travel method (Travel method)
Impact on: number of crashes, deaths, injuries.

var = 'Travel method'
df_rec.groupby(var).apply(basic).to_clipboard()
df_rec.groupby(var).apply(basic)

### 3.2 Road type (Road type)
Impact on: number of crashes, deaths, injuries.

var = 'Road type'
df_rec.groupby(var).apply(basic).to_clipboard()
df_rec.groupby(var).apply(basic)

### 3.3 Crash type (Crash type)
Impact on: number of crashes, deaths, injuries.

var = 'Crash type'
df_rec.groupby(var).apply(basic).to_clipboard()
df_rec.groupby(var).apply(basic)

### 3.4 Weather (Weather)
Impact on: number of crashes, deaths, injuries.

var = 'Weather'
df_rec.groupby(var).apply(basic).to_clipboard()
df_rec.groupby(var).apply(basic)

#### 3.4.1 Compare with the overall weather
Look at the weather records downloaded from OpenWeather.

df_weather = pd.read_csv(os.path.join(data_path, 'weather_shenzhen_2014-2016.csv'),
                         usecols=['dt_iso', 'weather_main'])
df_weather.dt_iso = df_weather.dt_iso.apply(lambda x: x.split(' ')[0])
df_weather.groupby('weather_main').size()

### 3.5 Crash causation (Crash causation)
Impact on: number of crashes, deaths, injuries.

var = 'Crash causation'
df_rec.groupby(var).apply(basic).to_clipboard()
df_rec.groupby(var).apply(basic)

### 3.6 Impact of month of year, day of week, and time of day
Impact on: number of crashes.

Run scr/plot_crash_by_time.R will get the plots.

### 3.7 Land-use cluster
Impact on: number of crashes.

var = 'Land-use cluster'
df_rec.groupby(var).apply(basic).to_clipboard()
df_rec.groupby(var).apply(basic)

# Check the road type of LUC2 vs LUC5
var1 = 'Land-use cluster'
var2 = 'Road type'
df_crs = df_rec.loc[:, [var1, var2]].pivot_table(index=var2, columns=var1, aggfunc=len, fill_value=0)
df_crs.drop(columns=['Unknown'], index=['Unknown'], inplace=True)
df_crs = df_crs/df_crs.sum(axis=0)*100

df_crs

# Check the deaths + injuries of LUC2 - rural high-speed vs LUC5 - urban low-speed
for var in ['Deaths', 'Injuries']:
    var1 = 'Land-use cluster'
    df_crs = df_rec.loc[df_rec[var1].isin(['LUC2', 'LUC5']), [var, var1]]
    conseq_LUC2 = np.mean(df_crs.loc[df_crs[var1] == 'LUC2', var].values)
    conseq_LUC5 = np.mean(df_crs.loc[df_crs[var1] == 'LUC5', var].values)

    print(f"{var}: LUC2 {conseq_LUC2}, LUC5 {conseq_LUC5}")
    print(stats.ks_2samp(df_crs.loc[df_crs[var1] == 'LUC2', var].values,
                         df_crs.loc[df_crs[var1] == 'LUC5', var].values))

### 3.8 Crash type x road type
Impact on: number of crashes. How distribution of crash type differs between different road types.

var1 = 'Crash type'
var2 = 'Road type'
df_crs = df_rec.loc[:, [var1, var2]].pivot_table(index=var2, columns=var1, aggfunc=len, fill_value=0)
df_crs.drop(columns=['Others', 'Unknown'], index=['Unknown'], inplace=True)
df_crs.to_csv(os.path.join(data_path, 'crash_x_road.csv'))

g, p, dof, expctd = stats.chi2_contingency(df_crs)
print(g, p, dof)
df_crs

crs_RT = df_crs.sum(axis=1)

### 3.9 Crash causation x Crash type, Crash causation x road type
Impact on: number of crashes. How distribution of crash type & road type differs between different Crash causation.

var1 = 'Crash causation'
var2 = 'Road type'
df_crs = df_rec.loc[:, [var1, var2]].pivot_table(index=var2, columns=var1, aggfunc=len, fill_value=0)
df_crs.drop(columns=['Others', 'Unknown'], index=['Unknown'], inplace=True)
df_crs.to_csv(os.path.join(data_path, 'causation_x_road.csv'))

g, p, dof, expctd = stats.chi2_contingency(df_crs)
print(g, p, dof)
df_crs

var1 = 'Crash causation'
var2 = 'Crash type'
df_crs = df_rec.loc[:, [var1, var2]].pivot_table(index=var2, columns=var1, aggfunc=len, fill_value=0)
df_crs.drop(columns=['Others', 'Unknown'], index=['Unknown', 'Others'], inplace=True)
df_crs.to_csv(os.path.join(data_path, 'causation_x_type.csv'))

g, p, dof, expctd = stats.chi2_contingency(df_crs)
print(g, p, dof)
df_crs

## 4 Statistics of different crash attributes: traffic participants
This part of the results are summarised in docs/article_tables.xlsx.
This part only focuses on the traffic participants who were driving a motor vehicle.

# Select motor-vehicle drivers
df = df.loc[df['Travel method'] == 'Driving motor vehicle', :]

print(f'Number of drivers involved in the crashes: {len(df)}')

tbl_head = ['Total', 'Full', 'Full%', 'Major', 'Major%', 'Equal', 'Equal%',
            'Minor', 'Minor%', 'No', 'No%', 'Unknown', 'Unknown%']

def res(data):
    mn_get = df_lvs.loc[df_lvs['var'] == 'Responsibility', ['en', 'abbr']]
    mn_get = mn_get.loc[mn_get['en'] != 'Unable to determine', :]
    data = pd.merge(data, mn_get,
                    left_on='Responsibility', right_on='abbr', how='left')
    num = data.groupby('Responsibility')['id'].size().reset_index()
    r = dict()
    r['Total'] = len(data) / len(df) * 100
    for _, row in num.iterrows():
        r[row['Responsibility']] = row['id']
        r[row['Responsibility']+ '%'] = row['id'] / len(data) * 100
    respon = pd.Series(r)
    for vr in tbl_head:
        if vr not in respon.index:
            respon[vr] = 0
    respon = respon[tbl_head]
    return respon

### 4.1 Gender of involved traffic participants
Impact on: responsibility.

df2test = df.groupby('Gender').apply(res)
df2test.drop(columns=['Total', 'Full%', 'Major%', 'Equal%', 'Minor%', 'No%', 'Unknown', 'Unknown%'],
             index=['Unknown'], inplace=True)
g, p, dof, _ = stats.chi2_contingency(df2test)
print(g, p, dof)

# Prepare data for the paper
df.groupby('Gender').apply(res).to_clipboard()
df.groupby('Gender').apply(res)

### 4.2 Age of involved traffic participants
Impact on: responsibility.

df2test = df.groupby('Age').apply(res)
df2test.drop(columns=['Total', 'Full%', 'Major%', 'Equal%', 'Minor%', 'No%', 'Unknown', 'Unknown%'],
             index=['Unknown'], inplace=True)
g, p, dof, _ = stats.chi2_contingency(df2test)
print(g, p, dof)

df.groupby('Age').apply(res).to_clipboard()
df.groupby('Age').apply(res)

### 4.3 Education level of involved traffic participants
Impact on: responsibility.

df.groupby('Education level').apply(res).to_clipboard()
df.groupby('Education level').apply(res)

### 4.4 Vehicle type of involved traffic participants
Impact on: responsibility.

df2test = df.groupby('Vehicle type').apply(res)
df2test.drop(columns=['Total', 'Full%', 'Major%', 'Equal%', 'Minor%', 'No%'], inplace=True)
g, p, dof, _ = stats.chi2_contingency(df2test)
print(g, p, dof)

df.groupby('Vehicle type').apply(res).to_clipboard()
df.groupby('Vehicle type').apply(res)

## 5 Prepare data for Bayesian network modelling
1 Focus on motor-vehicle drivers and their crashes like data used in the section 4 of this notebook.

2 Drop Education level because of too many missing values.

3 Merge injuries and deaths for the ease of interpretation.

4 Keep the crash records with clear information, i.e., no unknown and others fields.

# Drop Education level
df2bn = df.drop(columns=['Education level'])

# Merge injuries and deaths
var = 'Injuries and deaths'
df2bn.loc[:, var] = df2bn['Injuries'] + df2bn['Deaths']

# Check the distribution
df2bn_injdeaths = df2bn.groupby(var).size()
df2bn_injdeaths.name = 'count'
df2bn_injdeaths = df2bn_injdeaths.reset_index()
df2bn_injdeaths.loc[:, 'cum_count'] = df2bn_injdeaths['count'].cumsum()
df2bn_injdeaths.loc[:, 'cum_p'] = df2bn_injdeaths['count'].cumsum() / df2bn_injdeaths['count'].sum() * 100
df2bn_injdeaths.loc[:, 'p'] = df2bn_injdeaths['count'] / df2bn_injdeaths['count'].sum() * 100
df2bn_injdeaths.head(10)

df2bn.loc[:, var] = pd.cut(df2bn[var], bins=[-1, 0, 4, 9, 1000])
cat_dict = {'(-1, 0]': '= 0', '(9, 1000]': '> 9'}
df2bn.loc[:, var] = df2bn.loc[:, var].apply(lambda x: cat_dict[str(x)] if str(x) in cat_dict else str(x))
df2bn.loc[:, var].cat.add_categories("Unknown", inplace=True)
df2bn.loc[:, var].fillna("Unknown", inplace=True)

# Keep the crash records with complete information
df2bn.replace(to_replace = ['Unknown', 'Others'], value = np.nan, inplace = True)
df2bn.dropna(inplace=True)

### 5.1 Check crash type, driver responsibility -> injuries and deaths

df2bn.iloc[0]

var1 = 'Crash type'
var2 = 'Injuries and deaths'
df_crs = df2bn.loc[:, [var1, var2]].pivot_table(index=var2, columns=var1, aggfunc=len, fill_value=0)
df_crs = df_crs / df_crs.sum(axis=0)
df_crs.loc[['(4, 9]', '> 9'], :].sum()

var1 = 'Responsibility'
var2 = 'Injuries and deaths'
df_crs = df2bn.loc[:, [var1, var2]].pivot_table(index=var2, columns=var1, aggfunc=len, fill_value=0)
df_crs = df_crs / df_crs.sum(axis=0)
df_crs.loc[['(4, 9]', '> 9'], :].sum()

# Save it
print(f'Number of crashes/drivers to construct the Bayesian network: {len(df2bn)}')
df2bn.to_csv(os.path.join(data_path, 'data_s4.csv'), index=False)
